legacy: move pbkdf2 key file encryption to borg.legacy.crypto.key, refs #9556
